move compiler tools and minor edits
