+++ firedrake/slate/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import absolute_import, print_function, division
-from firedrake.slate.slate import *                     # noqa: F401
-from firedrake.slate.linear_algebra_compiler import *   # noqa: F401
+from firedrake.slate.slate import *  # noqa: F401
+from firedrake.slate.slac import *   # noqa: F401
