+++ firedrake/slate/slac/__init__.py
@@ -1,0 +1,2 @@
+from __future__ import absolute_import, print_function, division
+from firedrake.slate.slac.compiler import *  # noqa: F401
