+++ tests/regression/test_slate_infrastructure.py
@@ -226,12 +226,12 @@ def test_illegal_inverse():
 
 
 def test_illegal_compile():
-    from firedrake.slate import linear_algebra_compiler as slac
+    from firedrake.slate import compile_slate_expression as compile_slate
     V = FunctionSpace(UnitSquareMesh(1, 1), "CG", 1)
     v = TestFunction(V)
     form = v * dx
     with pytest.raises(ValueError):
-        slac.compile_slate_expression(form)
+        compile_slate(form)
 
 
 if __name__ == '__main__':
